pyproject.toml
[project]
name = "databox"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.11.11",
    "arrow>=1.3.0",
    "httpx>=0.28.1",
    "mitmproxy>=11.1.0",
    "peewee>=3.17.8",
    "playwright-stealth>=1.0.6",
    "pycryptodome>=3.21.0",
    "pydantic>=2.10.6",
    "pyexecjs>=1.5.1",
    "pymongo>=4.10.1",
    "qrcode>=8.0",
    "scrapy>=2.12.0",
    "scrapy-playwright>=0.0.43",
    "scrapy-redis>=0.9.1",
    "selenium>=4.28.1",
    "sqlalchemy>=2.0.38",
]
